Two More Functions
Added two more function stages which should be able to be completed relatively quickly. I don't have any of their logic in yet but once they're complete the hope is that I'll be able to start moving over to building the class object. Once the class(es) are done the processor.py function will be completed.