Base Format Done
The base format is done now need to reformat this into xsl, xml, asp type. Once that is done should be able to throw it to a hosting website and use the xml to store whatever variables we want. Might make a js updater to reformat the xml data each time the maps are updated.