Plotting Logic Finished
I've finished the plotting logic and now all thats left to do is get better interpolation and more data points. Once that's done the maps should be smoother and data more refined. Next step is to get the mesonet logic in so we can get the rest of our variables in.